+++ apps/web/src/state/info/hooks.ts
@@ -136,14 +136,20 @@ export const usePoolTransactionsSWR = (address: string): Transaction[] | undefin
 
 // Tokens hooks
 
-export const useAllTokenHighLight = (targetChainName?: MultiChainNameExtend): TokenData[] => {
+export const useAllTokenHighLight = ({
+  enable,
+  targetChainName,
+}: {
+  enable?: boolean
+  targetChainName?: MultiChainNameExtend
+}): TokenData[] => {
   const chainNameByQuery = useChainNameByQuery()
   const chainName = targetChainName ?? chainNameByQuery
   const [t24h, t48h, t7d, t14d] = getDeltaTimestamps()
   const { blocks } = useBlockFromTimeStampSWR([t24h, t48h, t7d, t14d], undefined, undefined, chainName)
   const type = checkIsStableSwap() ? 'stableSwap' : 'swap'
   const { data, isLoading } = useSWRImmutable(
-    blocks && chainName && [`info/token/data/${type}`, chainName],
+    enable && blocks && chainName && [`info/token/data/${type}`, chainName],
     () => fetchAllTokenData(chainName, blocks ?? []),
     SWR_SETTINGS_WITHOUT_REFETCH,
   )