+++ apps/web/src/views/FarmAuction/hooks/useWhitelistedAddresses.ts
@@ -1,13 +1,13 @@
-import useSWR from 'swr'
 import { useFarmAuctionContract } from 'hooks/useContract'
 import { getBidderInfo } from 'config/constants/farmAuctions'
 import { FarmAuctionBidderConfig } from 'config/constants/types'
 import { AUCTION_WHITELISTED_BIDDERS_TO_FETCH } from 'config'
+import { useQuery } from '@tanstack/react-query'
 
 const useWhitelistedAddresses = (): FarmAuctionBidderConfig[] => {
   const farmAuctionContract = useFarmAuctionContract()
 
-  const { data } = useSWR(['farmAuction', 'whitelistedAddresses'], async () => {
+  const { data } = useQuery(['farmAuction', 'whitelistedAddresses'], async () => {
     try {
       const [bidderAddresses] = await farmAuctionContract.read.viewBidders([
         0n,