+++ packages/chains/src/index.test.ts
@@ -6,7 +6,13 @@ test('exports', () => {
     [
       "ChainId",
       "chainNames",
+      "chainNameToChainId",
+      "defiLlamaChainNames",
       "getChainName",
+      "getLlamaChainName",
+      "getChainIdByChainName",
+      "V3_SUBGRAPHS",
+      "V2_SUBGRAPHS",
     ]
   `)
 })